moved backend code back into server, add . before relative file imports in app.py and models.py
