+++ .gitignore
@@ -1,4 +1,4 @@
-.env
+server/.env
 
 #amplify-do-not-edit-begin
 amplify/\#current-cloud-backend
