+++ Procfile
@@ -1,1 +1,1 @@
-web: gunicorn -w 4 -b 0.0.0.0:$PORT app:app
+web: gunicorn -w 4 -b 0.0.0.0:$PORT server.app:app
