+++ server/alembic/env.py
@@ -3,7 +3,7 @@ from sqlalchemy import engine_from_config
 from sqlalchemy import pool
 from alembic import context
 import os
-import models
+import server.models as models
